Allow `litra-autotoggle` to run when there is no Litra device connected
At the moment, `litra-autotoggle` exits with an error if no
Litra device is connected. This is annoying if you want to run
in the background, but you don't always have a Litra connected
(e.g. you sometimes unplug it!).

This switches to a new approach where, by default, the command
will keep on running even if a Litra device is not connected,
but it will be a no-op.

Optionally, the `--require-device` argument allows you to enforce
that a device is connected, exiting if it isn't.

Fixes #2.